requirements.txt
autopep8==1.5.7
numpy==1.21.1
Pillow==8.3.1
pycodestyle==2.7.0
toml==0.10.2
